faculty service

+++ src/main/java/io/bgnc/iytechsocialmediaapplication/service/CommunityService.java
@@ -5,7 +5,7 @@ import io.bgnc.iytechsocialmediaapplication.model.HeadOfCommunity;
 import io.bgnc.iytechsocialmediaapplication.repository.CommunityRepository;
 import io.bgnc.iytechsocialmediaapplication.repository.HeadOfCommunityRepository;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -16,11 +16,22 @@ import java.util.List;
 @Transactional
 public class CommunityService {
 
-    @Autowired
-    private CommunityRepository communityRepository;
 
-    @Autowired
-    private HeadOfCommunityRepository headOfCommunityRepository;
+    private final CommunityRepository communityRepository;
+
+
+    private final HeadOfCommunityRepository headOfCommunityRepository;
+
+
+    public CommunityService(CommunityRepository communityRepository,
+                            HeadOfCommunityRepository headOfCommunityRepository)
+    {
+        this.communityRepository = communityRepository;
+        this.headOfCommunityRepository = headOfCommunityRepository;
+    }
+
+
+
 
 
     public Community saveCommunity(Community community) throws IllegalAccessException {
@@ -53,7 +64,21 @@ public class CommunityService {
 
     public void deleteCommunity(Long community_id) {
 
-        communityRepository.deleteById(community_id);
+        List<Community> allCommunities = getAllCommunities();
+
+        if(communityRepository.existsById(community_id)){
+
+            for (int i = 0 ; i < allCommunities.size();i++){
+                if (headOfCommunityRepository.existsById(communityRepository.findById(community_id).get().getHeadOfCommunity_id().get(i).getHead_of_community_id())) {
+                    communityRepository.deleteById(community_id);
+                }
+                else
+                    log.warn("Head Of Community is not found ");
+            }
+
+
+        }
+
     }
 
     public Community updateCommunity(Long community_id, String headOfCommunityName,String community_name,String community_photo,String community_description){
